EZP-29802: Inline custom tags support
