+++ src/bundle/Resources/public/js/alloyeditor/src/buttons/ez-btn-customtag-update.js
@@ -16,6 +16,12 @@ export default class EzBtnCustomTagUpdate extends EzWidgetButton {
         };
     }
 
+    componentDidMount() {
+        if (!Object.keys(this.attributes).length) {
+            this.saveCustomTag();
+        }
+    }
+
     /**
      * Renders the text input.
      *
